requirements.txt
streamlit==1.29.0
llama-index==0.9.19
langchain==0.0.352
python-dotenv==1.0.0
cryptography==41.0.7
pypdf==3.17.3
watchdog==3.0.0